Group04/Archive_Folder/Deep_Research_V00/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
pydantic>=2.0.0,<3.0.0
tavily-python>=0.3.0
duckduckgo-search>=4.1.0
pypdf>=3.17.0
python-docx>=1.1.0
reportlab>=4.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
python-dotenv>=1.0.0
pillow>=10.0.0
watchdog>=3.0.0
wikipedia>=1.4.0
arxiv>=2.1.0
sentence-transformers>=2.2.0
xmltodict>=0.13.0
